# openapi.yaml
openapi: 3.0.3
info:
  title: Tereka Online API
  version: "1.0.0"
  description: >
    Multi-SACCO management platform API for Uganda. This is a curated specification of the
    integrator-facing surface (authentication, member self-service, mobile-money integration, and the
    core financial/loan read + approval endpoints). It is maintained by hand alongside the code; the
    authoritative behaviour is the Spring Boot backend under `backend-java`.


    ## Conventions

    - Base path: `/api/v1`.

    - Successful responses are wrapped in an envelope: `{ "data": <payload> }`.

    - Errors use `{ "error": { "timestamp", "status", "code", "message" } }`.

    - Authentication is a bearer token obtained from a login endpoint, sent as
      `Authorization: Bearer <token>`.

    - Abuse-prone endpoints are rate limited per client IP and may return `429` with a `Retry-After`
      header.
  contact:
    name: Tereka Online
servers:
  - url: "https://{host}/api/v1"
    description: Deployment
    variables:
      host:
        default: tereka.online
  - url: "http://127.0.0.1:8080/api/v1"
    description: Local backend
...
security:
  - bearerAuth: []
paths:
...
  /fund-types:
    get:
      tags: [Accounting]
      summary: List the SACCO configurable fund types (fund-types:view)
      parameters:
        - { name: tenantId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Fund types
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/FundType" } }
        "403": { $ref: "#/components/responses/Error" }
    post:
      tags: [Accounting]
      summary: Add a SACCO fund type (fund-types:manage)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FundTypeRequest" }
      responses:
        "201":
          description: Created fund type
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/FundType" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /fund-types/{fundTypeId}:
    patch:
      tags: [Accounting]
      summary: Edit a SACCO fund type (fund-types:manage)
      parameters:
        - { name: fundTypeId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FundTypeRequest" }
      responses:
        "200":
          description: Updated fund type
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/FundType" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Staff/platform session token from `/auth/login`.
    memberBearer:
      type: http
      scheme: bearer
      description: Member session token from `/member-auth/login`.
    callbackSignature:
      type: apiKey
      in: header
      name: X-Mobile-Money-Signature
      description: HMAC-SHA256 of the raw request body using the shared callback secret.
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    FundType:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        code: { type: string, example: burial }
        name: { type: string, example: Burial Fund }
        basis: { type: string, description: savings | shares | welfare }
        description: { type: string, nullable: true }
        system: { type: boolean, description: true for built-in Savings/Shares/Welfare }
        active: { type: boolean }
        displayOrder: { type: integer }
        createdByUserId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    FundTypeRequest:
      type: object
      required: [name]
      properties:
        tenantId: { type: string, nullable: true }
        code: { type: string, description: "lowercase; required on create, ignored on edit" }
        name: { type: string }
        basis: { type: string, description: savings | shares | welfare }
        description: { type: string, nullable: true }
        active: { type: boolean, nullable: true }
        displayOrder: { type: integer, nullable: true }
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            timestamp: { type: string, format: date-time }
            status: { type: integer, example: 400 }
            code: { type: string, example: "AUTH_INVALID" }
            message: { type: string }
...